requirements.txt
SQLAlchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
fastapi==0.120.0
greenlet==3.2.4
psycopg2-binary==2.9.11
pydantic-settings=2.11.0
pydantic==2.12.3
python-telegram-bot==22.5
uvicorn==0.38.0